Move current routers to the 'old' namespace
This is to indicate that they purely exist for backwards compatibility
