+++ api/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from starlette.middleware.authentication import AuthenticationMiddleware
 
 from api.middleware import TokenAuthentication, on_auth_error
-from api.routers import codejams, infractions, teams, users, winners
+from api.routers.old import old_routes_router
 from api.settings import Server
 
 app = FastAPI(redoc_url="/", docs_url="/swagger")
@@ -13,8 +13,4 @@ app.add_middleware(
     on_error=on_auth_error,
 )
 
-app.include_router(codejams.router)
-app.include_router(infractions.router)
-app.include_router(teams.router)
-app.include_router(users.router)
-app.include_router(winners.router)
+app.include_router(old_routes_router)
